+++ akka-guard-core-typed/src/main/scala/com/chatwork/akka/guard/typed/config/SABConfig.scala
@@ -1,4 +1,4 @@
-package com.chatwork.akka.guard.typed
+package com.chatwork.akka.guard.typed.config
 
 import scala.concurrent.duration.{ Duration, FiniteDuration }
